Move files to prototype/sparsity
