+++ torchao/prototype/sparsity/pruner/images/prune_1.png
