+++ torchao/prototype/sparsity/pruner/images/prune_2.png
