+++ torchao/prototype/sparsity/pruner/images/prune_4.png
