+++ torchao/prototype/sparsity/pruner/images/prune_6.png
